requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.2
